more complete unit tests

+++ internal/oplog/writer.go
@@ -1,6 +1,7 @@
 package oplog
 
 import (
+	"errors"
 	"fmt"
 	"reflect"
 	"strings"
@@ -27,6 +28,12 @@ type GormWriter struct {
 
 // Create an object in storage
 func (w *GormWriter) Create(i interface{}) error {
+	if w.Tx == nil {
+		return errors.New("create Tx is nil")
+	}
+	if i == nil {
+		return errors.New("create interface is nil")
+	}
 	if err := w.Tx.Create(i).Error; err != nil {
 		return fmt.Errorf("error creating: %w", err)
 	}
@@ -36,6 +43,12 @@ func (w *GormWriter) Create(i interface{}) error {
 // Update an object in storage, if there's a fieldMask then only the field_mask.proto paths are updated, otherwise
 // we will send every field to the DB.
 func (w *GormWriter) Update(i interface{}, fieldMaskPaths []string) error {
+	if w.Tx == nil {
+		return errors.New("update Tx is nil")
+	}
+	if i == nil {
+		return errors.New("update interface is nil")
+	}
 	if len(fieldMaskPaths) == 0 {
 		if err := w.Tx.Save(i).Error; err != nil {
 			return fmt.Errorf("error updating: %w", err)
@@ -50,13 +63,16 @@ func (w *GormWriter) Update(i interface{}, fieldMaskPaths []string) error {
 			// support for an embedded a gorm type
 			if structTyp.Field(i).Type.Kind() == reflect.Struct {
 				embType := structTyp.Field(i).Type
-				embVal := reflect.Indirect(reflect.ValueOf(val.Field(i).Interface()))
-				for embFieldNum := 0; embFieldNum < embType.NumField(); embFieldNum++ {
-					if strings.EqualFold(embType.Field(embFieldNum).Name, field) {
-						updateFields[field] = embVal.Field(embFieldNum).Interface()
+				// check if the embedded field is exported via CanInterface()
+				if val.Field(i).CanInterface() {
+					embVal := reflect.Indirect(reflect.ValueOf(val.Field(i).Interface()))
+					for embFieldNum := 0; embFieldNum < embType.NumField(); embFieldNum++ {
+						if strings.EqualFold(embType.Field(embFieldNum).Name, field) {
+							updateFields[field] = embVal.Field(embFieldNum).Interface()
+						}
 					}
+					continue
 				}
-				continue
 			}
 			// it's not an embedded type, so check if the field name matches
 			if strings.EqualFold(structTyp.Field(i).Name, field) {
@@ -72,6 +88,12 @@ func (w *GormWriter) Update(i interface{}, fieldMaskPaths []string) error {
 
 // Deleting an object in storage
 func (w *GormWriter) Delete(i interface{}) error {
+	if w.Tx == nil {
+		return errors.New("delete Tx is nil")
+	}
+	if i == nil {
+		return errors.New("delete interface is nil")
+	}
 	if err := w.Tx.Delete(i).Error; err != nil {
 		return fmt.Errorf("error deleting: %w", err)
 	}
